+++ R/elevation.R
@@ -77,7 +77,7 @@ format_out <- function(x) {
 #'
 #' @return a matrix, or depending on format of 'extent' a terra rast or raster object with elevation data
 #' @export
-#' @importFrom dsn cop30 cop90
+#' @importFrom spatial.datasources cop30 cop90
 #' @examples
 #'
 #' image(elevation(), useRaster = TRUE)
@@ -116,7 +116,7 @@ elevation <- function(extent = c(-180, 180, -90, 90), ..., dimension = NULL, pro
 
   if (is.null(source)) {
     rso <- c(gebco(),
-             dsn::cop90(), dsn::cop30())
+             spatial.datasources::cop90(), spatial.datasources::cop30())
     if (no_srtm30) rso <- rso[1:2] else rso <- rso[c(1, 3)]
     if (no_srtm) rso <- rso[1L]
     if (!no_srtm) print("SRTM in use, in addition to GEBCO")